refactor(#1731,#1732,#1733): delete old agent_runtime, move agent_loop to agent_runtime/loop
- Delete obsolete agent_runtime/{process_manager,session_store,types}.py
  (superseded by core/process_table.py + contracts/process_types.py)
- Delete contracts/protocols/process_manager.py (no consumers)
- Move system_services/agent_loop.py → agent_runtime/loop.py
  (shared base for managed + unmanaged agent loops)
- Remove AgentRuntimePlaceholder + boot block from factory/_system.py
- Update all imports and test expected_keys
