+++ src/nexus/factory/_system.py
@@ -17,26 +17,11 @@ Two severity classes:
 import logging
 import time
 from collections.abc import Callable
-from dataclasses import dataclass
 from typing import Any, cast
 
 from nexus.contracts.constants import ROOT_ZONE_ID
 from nexus.factory._boot_context import _BootContext
 from nexus.factory._helpers import _make_gate
-
-
-@dataclass(frozen=True, slots=True)
-class AgentRuntimePlaceholder:
-    """Two-phase wiring placeholder for agent runtime.
-
-    Created during system-service boot (Phase 1) with partial dependencies.
-    Fully wired by orchestrator.py (Phase 2) once NexusFS + LLM are available.
-    """
-
-    factory_class: type
-    sandbox: Any
-    scheduler: Any
-
 
 logger = logging.getLogger(__name__)
 
@@ -518,27 +503,6 @@ def _boot_system_services(
         except Exception as exc:
             logger.warning("[BOOT:SYSTEM] AcpService unavailable: %s", exc)
 
-    # --- Agent Runtime (Agent Process Engine, AGENT-PROCESS-ARCHITECTURE) ---
-    agent_runtime: Any = None
-    if _on("agent_runtime"):
-        try:
-            from nexus.system_services.agent_runtime.process_manager import (
-                ProcessManager as _AgentProcessManager,
-            )
-
-            # LLM provider is wired later (in _boot_wired_services) since
-            # it needs NexusFS.  For now, agent_runtime is constructed lazily
-            # at first use via the two-phase pattern.  Store the class + deps
-            # so orchestrator.py can finish wiring after NexusFS is created.
-            agent_runtime = AgentRuntimePlaceholder(
-                factory_class=_AgentProcessManager,
-                sandbox=None,  # wired later
-                scheduler=scheduler_service,
-            )
-            logger.debug("[BOOT:SYSTEM] AgentRuntime placeholder created (two-phase)")
-        except Exception as exc:
-            logger.warning("[BOOT:SYSTEM] AgentRuntime unavailable: %s", exc)
-
     # =====================================================================
     # Assemble result
     # =====================================================================
@@ -571,7 +535,6 @@ def _boot_system_services(
         "zone_lifecycle": zone_lifecycle,
         "scheduler_service": scheduler_service,
         "acp_service": acp_service,
-        "agent_runtime": agent_runtime,
     }
 
     elapsed = time.perf_counter() - t0
