+++ src/nexus/system_services/agent_runtime/loop.py
